Major Release Added and removed a lot of things.
Major additions include PHP 8.1, OpenSSL 3.0, and MariaDB 10.7.

A bug in WP-CLI was preventing it from working with PHP 8.1. A bug 
report was submitted to the WP-CLI team and they fixed the issue in 
recent weeks. EngineScript uses WP-CLI to create new WordPress 
installations, and this bug was holding up further code releases.

It is recommended to test this major release on a fresh server. Many 
things were changed over the past months and upgrade scripts will likely 
not be able to account for all changes. Moving forward, upgrade scripts 
will be better fleshed out and releases should flow better now that the 
WP-CLI bug with PHP 8.1 has been fixed.