Removed Super Page Cache
This plugin has moved in a very different direction that adds a lot of 
complexity and could cause major issues with EngineScript if the plugin 
is misconfigured. 

In addition, the latest versions have added a ton of bugs and bulk to 
the plugin. They also inject advertisements into a few places in the 
wordpress dashboard that cannot be disabled.