Removed PHP Extensions
Removed a few PHP extensions that are not regularly used by WordPress.

One exception was SOAP. I was able to locate a non-wordpress.org plugin 
for Amazon product CLI that required SOAP. I'm not sure how widely this 
plugin is used at all, and as such I don't think SOAP is a must have 
bundle for default installations.

Readline is not really widely used at all either, but I was able to find 
enough small uses of it to warrant leaving it in.