ACME.sh ZeroSSL
It's been a while since ACME.sh switched to using ZeroSSL and it's high 
time that we switched.

Let's Encrypt is discontinuing use of OCSP, while ZeroSSL still does. So 
it's a very good time to switch.

Registration is done in the script, no need for the user to register for 
an account.