Prerelease End of Life
EngineScript began its life as a smaller in-house project in 2016. In 
2018 it officially transitioned into a separate project and has been in 
various degrees of development since then. In 2025 we plan on moving 
through alpha, beta, and final release stages.

With that being said, some servers could be running highly out of date 
versions of EngineScript that make it next to impossible to bring up to 
date through auto-upgrade sed scripts. This commit kills off the 
auto-upgrade cron and prevents it from being added again on legacy 
servers that are using an older user config file that doesn't include 
the new variable. It's not a perfect solution, but it's likely the only 
one available that doesn't break servers. We already did that 
accidentally a few weeks ago.

So, if you've been using EngineScript and enjoy it, please consider 
spinning up a new server and do a fresh install on EngineScript.

If you need some advice on transfering your old site to a new server, 
please contact me privately at peter@visistruct.com.