PHP Tuning
rewrote the PHP tuning scripts to simplify the tuning process. Also moved from ondemand to dynamic as it will likely improve TTFB on sites with higher traffic.

Dropbox uploader is likely abandoned and is not as intuitive as amazon's s3 backup software is, so dropbox support is removed until a better solution comes along.