Updates
Updated nginx to better handle woocoomerce caching errors. 

A trello issue popped up where users were encountering a redirrect loop issue that could occur under unknown, conditions. This issue has been observed before in some complex nginx installations that use caching. I've personally encountered this issue before as well but was never able to reproduce it reliably. This update adds some nginx configuration changes that could possibly help mitigate the issue, as it may be widespread across many other nginx projects but hasn't been reported yet. It's quite possible that enginescript also suffers from this issue under very specific conditions. 

My theory is that wordpress' cron function is triggering a daily update or health check at the same time a user is trying to access the site, and if a bot user attempts to view the site (possibly without SSL) while the update is occurring, it could trigger a redirect loop. This is just a theory, but the changes made in this update should help mitigate this issue if it does occur.